fix: 빌드에러 처리용 코드분리
