+++ src/app/(life)/life/page.tsx
@@ -4,16 +4,9 @@ import Calendar from '@/components/features/life/calendar';
 import SoloLifeList from '@/components/features/life/solo-life-list';
 import { PATH } from '@/constants/page-path';
 import { useLifePostsByMonth } from '@/lib/hooks/queries/use-life-posts-by-month';
+import { getToday } from '@/lib/utils/get-today';
 import Link from 'next/link';
 import { useMemo, useState } from 'react';
-
-export const getToday = (): string => {
-  const kst = new Date();
-  const yyyy = kst.getFullYear();
-  const mm = String(kst.getMonth() + 1).padStart(2, '0');
-  const dd = String(kst.getDate()).padStart(2, '0');
-  return `${yyyy}-${mm}-${dd}`;
-};
 
 const LifePage = () => {
   const [selectedDate, setSelectedDate] = useState<string>(getToday());
