+++ src/components/common/post-input-form.tsx
@@ -10,8 +10,8 @@ import { supabase } from '@/lib/utils/supabase/supabase-client';
 import { useLifePost } from '@/lib/hooks/mutations/useLifePost';
 import { v4 as uuidv4 } from 'uuid';
 import { PATH } from '@/constants/page-path';
-import { getToday } from '@/app/(life)/life/page';
 import DatePicker from './date-picker';
+import { getToday } from '@/lib/utils/get-today';
 
 type LifeInputFormProps = {
   userId: string;
