+++ src/lib/utils/get-today.ts
@@ -1,0 +1,7 @@
+export const getToday = (): string => {
+  const kst = new Date();
+  const yyyy = kst.getFullYear();
+  const mm = String(kst.getMonth() + 1).padStart(2, '0');
+  const dd = String(kst.getDate()).padStart(2, '0');
+  return `${yyyy}-${mm}-${dd}`;
+};
